fixed obj_affect started on new vehicle code
git-svn-id: svn+ssh://4dimensions.org/home/game-source/repositories/gamesrc@707 254f5b4c-bdf4-4503-907e-934c7b59e499

+++ structs.h
@@ -1502,7 +1502,7 @@ struct obj_flag_data {
 /* Used in obj_file_elem *DO*NOT*CHANGE* */
 struct obj_affected_type {
     byte location;       /* Which ability to change (APPLY_XXX) */
-    sbyte modifier;      /* How much it changes by              */
+    short int modifier;      /* How much it changes by              */
 };
